+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - "Abbreviations" have been renamed to "Callsigns"
 - There are now badges in transmitted messages to define admins, server owners, and the room owner
 - Renamed "Management Role" to "Admin Role"
+- The announce button has been turned into a command
 
 ### **Bug Fixes**
 - Unauthorized users can no longer close polls